LP64: fix the wait-queue hash, an ipip.c uninit read, and a dead TTL sentinel
phash() (ksubr.c) is the worst of the three.  On the !HAS_UCONTEXT path -
which is the one macOS builds, configure reports HAS_UCONTEXT 0 - it read

    return (int)event % PHASH;

and the function returns unsigned.  Truncating the event pointer to int makes
it negative whenever bit 31 of the low half is set; C99 then gives a negative
remainder, and returning that through an unsigned result type turns it into a
huge index.  kernel.c uses it to both read and write Waittab[PHASH], a
16-element array:

    Stack   0x7ff7ba02ebf8   old=4294967288   new=14   <<< 4.29e9 into [16]
    BSS     0x105ed2000      old=0            new=0
    Heap    0x600001830000   old=0            new=0

An event on the stack that is not 16-byte aligned therefore indexes about
32 GiB past Waittab.

Why it never fired: WAMPES only ever passes globals and heap objects as
events - NULL, &Killq, &Hopper, &iface->outq, Curproc, a queue head - never a
stack address.  On macOS those land at 0x1xxxxxxxx (PIE image) and
0x6000xxxxxxxx (malloc magazine), and the low 32 bits of both stay below
0x80000000, so the int comes out positive.  Sampling 1800 such addresses over
300 runs produced no index out of range.  So it survived on the shape of the
memory map, not by design: an mmap-backed allocation, or anyone using a local
as an event, would hit it immediately.

On Linux the addresses would be far less accommodating - PIE heap around
0x55xxxxxxxxxx, mmap at 0x7fxxxxxxxxxx - but there HAS_UCONTEXT is 1 and the
other branch already computed ((long)event >> 2) % PHASH, which is correct.
Two accidents cancelling out.

Now computed in unsigned long for both paths, so the #if goes away too.

ipip_receive(): in UDP mode "l <= 0" was the only length check, but a few
lines further down get32(bufptr + 12) reads the source address of the
encapsulated IP header.  A one-byte datagram therefore read four bytes that
were never received - uninitialised stack - and rt_add() put the result in
the IP routing table.  Now requires a full IP header.  Same int-against-
size_t comparison as axip.c had in 77a741e, and addrlen is socklen_t.
"attach ipip" is in the shipped net.rc, next to "attach axip".

TTL_MISSING was 0x80000000UL compared against an int32, so on LP64 the int32
sign-extends and the sentinel can never match - clang flags it on every
build.  Nothing assigns TTL_MISSING anywhere, so the behaviour does not
change; the macro is simply typed correctly now.

Also in pproc(): the "in"/"out" columns printed a FILE * truncated to int.
They show fileno() now.

src, bbs and tools build; the only warnings left in src are two
-Wstring-plus-int on the legitimate month-name table idiom, the deprecated
utmp on macOS, the crypt() prototype (open, W-35) and the constant-string
-Wformat-security noise.

+++ src/domain.h
@@ -124,7 +124,11 @@ struct rr {
 	char *comment;          /* optional comment */
 	char *name;             /* Domain name, ascii form */
 	int32 ttl;              /* Time-to-live */
-#define TTL_MISSING     0x80000000UL
+/* Cast matters: ttl is int32.  Against a bare 0x80000000UL the int32 is
+ * widened to unsigned long, so on LP64 the sentinel value sign-extends to
+ * 0xFFFFFFFF80000000 and the comparison in put_rr() can never match.
+ */
+#define TTL_MISSING     ((int32) 0x80000000UL)
 	uint class;             /* IN, etc */
 #define CLASS_MISSING   0
 	uint type;              /* A, MX, etc */
